# [Xamarin.Android.Build.Tasks] SRM in <GetAdditionalResourcesFromAssemblies/>