Just like in c22475d, we can use System.Reflection.Metadata in places
over Mono.Cecil for performance benefits.

On such case is the `<GetAdditionalResourcesFromAssemblies/>` MSBuild
task, which only looks at custom attributes in assemblies:

    Before
    1st:          98 ms  GetAdditionalResourcesFromAssemblies       1 calls
    Incremental: 105 ms  GetAdditionalResourcesFromAssemblies       1 calls

    After
    1st:          54 ms  GetAdditionalResourcesFromAssemblies       1 calls
    Incremental:  53 ms  GetAdditionalResourcesFromAssemblies       1 calls

This was a straightforward port to SRM, which seems to save ~50ms on
every build.